Added some icicle wrappers fully support icicle functionality needed by gnark, cleaned up
